authorization task by getUser() decorators

+++ src/tasks/tasks.service.ts
@@ -1,3 +1,4 @@
+import { User } from 'src/auth/user.entity';
 import { GetTaskFilterDto } from './dto/get-tasks-filter.dto';
 import { CreateTaskDto } from './dto/create-task.dto';
 
@@ -14,17 +15,20 @@ export class TasksService {
   }
   // private tasks: Task[] = [];
 
-  async getTasks(filterDto: GetTaskFilterDto): Promise<Task[]> {
-    return await this.taskRepository.getTasks(filterDto)
+  async getTasks(filterDto: GetTaskFilterDto,
+    user: User): Promise<Task[]> {
+    return await this.taskRepository.getTasks(filterDto, user)
   }
   // async getTasks(): Promise<Task[]> {
   //   return await this.taskRepository.find();
   // }
-  async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-    return this.taskRepository.createTask(createTaskDto)
+  async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+    return this.taskRepository.createTask(createTaskDto, user)
   }
-  async getTaskById(id: number): Promise<Task> /*Task for entity*/ {
-    const found = await this.taskRepository.findOne(id);
+  async getTaskById(id: number, user: User): Promise<Task> /*Task for entity*/ {
+    const found = await this.taskRepository.findOne({
+      where: { id, userId: user.id }
+    });
     if (!found) {
       throw new NotFoundException(`This id ${id} is doesn't exist`);
     }
@@ -32,14 +36,14 @@ export class TasksService {
   }
 
 
-  async deleteTaskById(id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id);
+  async deleteTaskById(id: number, user: User): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id });
     if (!result.affected) {
       throw new NotFoundException(`This id ${id} is doesn't exist`);
     }
   }
-  async updateStatusTaskById(id: number, status: TaskStatus): Promise<Task> {
-    const task: Task = await this.getTaskById(id);
+  async updateStatusTaskById(id: number, status: TaskStatus, user: User): Promise<Task> {
+    const task: Task = await this.getTaskById(id, user);
     task.status = status;
     await task.save()
     return task;
